# openapi/coinbase-admin-api.yaml
openapi: 3.1.0
info:
  title: Coinbase Admin API
  description: Enterprise API for the Coinbase trading engine. Read-only operator
    routes are active; mutating HTTP routes use shared command services but remain
    live-disabled by the approval gate.
  version: 0.1.0
paths:
...
  /api/v1/spot/readiness:
    get:
      tags:
      - spot
      summary: Read spot trading readiness
      operationId: spot_readiness_api_v1_spot_readiness_get
      parameters:
      - name: product_id
        in: query
        required: false
        schema:
          anyOf:
          - type: array
            items:
              type: string
          - type: 'null'
          title: Product Id
      - name: Authorization
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-Admin-Actor
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Actor
      - name: X-Admin-Roles
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Roles
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpotReadinessResponse'
        '401':
          description: Missing or invalid Admin API authentication.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiErrorResponse'
        '403':
          description: Actor lacks the required Admin API permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/spot/sweep/status:
    get:
      tags:
      - spot
      summary: Read spot sweep status
      operationId: spot_sweep_status_api_v1_spot_sweep_status_get
      parameters:
      - name: state_file
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: State File
      - name: Authorization
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-Admin-Actor
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Actor
      - name: X-Admin-Roles
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Roles
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpotSweepStatusResponse'
        '401':
          description: Missing or invalid Admin API authentication.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiErrorResponse'
        '403':
          description: Actor lacks the required Admin API permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/spot/sweep/pnl:
    get:
      tags:
      - spot
      summary: Read spot sweep P/L
      operationId: spot_sweep_pnl_api_v1_spot_sweep_pnl_get
      parameters:
      - name: product_id
        in: query
        required: false
        schema:
          anyOf:
          - type: array
            items:
              type: string
          - type: 'null'
          title: Product Id
      - name: include_coinbase_average_cost
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Include Coinbase Average Cost
      - name: Authorization
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-Admin-Actor
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Actor
      - name: X-Admin-Roles
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Roles
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpotSweepPnlResponse'
        '401':
          description: Missing or invalid Admin API authentication.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiErrorResponse'
        '403':
          description: Actor lacks the required Admin API permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/spot/cost-basis/status:
    get:
      tags:
      - spot
      summary: Read spot cost-basis status
      operationId: spot_cost_basis_status_api_v1_spot_cost_basis_status_get
      parameters:
      - name: state_file
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: State File
      - name: Authorization
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-Admin-Actor
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Actor
      - name: X-Admin-Roles
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Roles
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpotCostBasisStatusResponse'
        '401':
          description: Missing or invalid Admin API authentication.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiErrorResponse'
        '403':
          description: Actor lacks the required Admin API permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/spot/campaign/status:
    get:
      tags:
      - spot
      summary: Read spot campaign status
      operationId: spot_campaign_status_api_v1_spot_campaign_status_get
      parameters:
      - name: state_file
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: State File
      - name: Authorization
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-Admin-Actor
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Actor
      - name: X-Admin-Roles
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Roles
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpotCampaignStatusResponse'
        '401':
          description: Missing or invalid Admin API authentication.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiErrorResponse'
        '403':
          description: Actor lacks the required Admin API permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/spot/direct-orders/{client_order_id}/audit:
    get:
      tags:
      - spot
      summary: Read direct spot order audit by client_order_id
      operationId: spot_direct_order_audit_api_v1_spot_direct_orders__client_order_id__audit_get
      parameters:
      - name: client_order_id
        in: path
        required: true
        schema:
          type: string
          title: Client Order Id
      - name: include_events
        in: query
        required: false
        schema:
          type: boolean
          default: true
          title: Include Events
      - name: include_fills
        in: query
        required: false
        schema:
          type: boolean
          default: true
          title: Include Fills
      - name: event_limit
        in: query
        required: false
        schema:
          type: integer
          default: 100
          title: Event Limit
      - name: fill_limit
        in: query
        required: false
        schema:
          type: integer
          default: 1000
          title: Fill Limit
      - name: Authorization
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      - name: X-Admin-Actor
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Actor
      - name: X-Admin-Roles
        in: header
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Admin-Roles
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpotDirectOrderAuditResponse'
        '401':
          description: Missing or invalid Admin API authentication.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiErrorResponse'
        '403':
          description: Actor lacks the required Admin API permission.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminApiErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    AdminApiErrorCode:
      type: string
      enum:
      - auth_required
      - permission_denied
      - validation_error
      - idempotency_conflict
      - guard_blocked
      - not_implemented
      - backend_unavailable
      - request_error
      title: AdminApiErrorCode
      description: Structured Admin API error codes exposed to the frontend.
    AdminApiErrorResponse:
      properties:
        code:
          $ref: '#/components/schemas/AdminApiErrorCode'
        message:
          type: string
          title: Message
        severity:
          $ref: '#/components/schemas/AdminApiErrorSeverity'
        guard_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Guard Name
        field_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Field Path
        correlation_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Correlation Id
        audit_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Audit Id
        live_coinbase_orders_ran:
          type: boolean
          title: Live Coinbase Orders Ran
          default: false
      additionalProperties: false
      type: object
      required:
      - code
      - message
      - severity
      title: AdminApiErrorResponse
      description: Structured error body shared by Admin API routes.
    AdminApiErrorSeverity:
      type: string
      enum:
      - info
      - warning
      - error
      - critical
      title: AdminApiErrorSeverity
      description: Operator-facing severity for structured Admin API errors.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    SpotCampaignStatusResponse:
      properties:
        type:
          anyOf:
          - type: string
          - type: 'null'
          title: Type
        status:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
        live_coinbase_orders_ran:
          type: boolean
          title: Live Coinbase Orders Ran
          default: false
      additionalProperties: true
      type: object
      title: SpotCampaignStatusResponse
      description: Spot campaign status response.
    SpotCostBasisStatusResponse:
      properties:
        type:
          anyOf:
          - type: string
          - type: 'null'
          title: Type
        status:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
        live_coinbase_orders_ran:
          type: boolean
          title: Live Coinbase Orders Ran
          default: false
      additionalProperties: true
      type: object
      title: SpotCostBasisStatusResponse
      description: Spot cost-basis status response.
    SpotDirectOrderAuditResponse:
      properties:
        type:
          anyOf:
          - type: string
          - type: 'null'
          title: Type
        status:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
        live_coinbase_orders_ran:
          type: boolean
          title: Live Coinbase Orders Ran
          default: false
        client_order_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Client Order Id
      additionalProperties: true
      type: object
      title: SpotDirectOrderAuditResponse
      description: Direct spot order audit response keyed by ``client_order_id``.
    SpotReadinessResponse:
      properties:
        type:
          anyOf:
          - type: string
          - type: 'null'
          title: Type
        status:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
        live_coinbase_orders_ran:
          type: boolean
          title: Live Coinbase Orders Ran
          default: false
      additionalProperties: true
      type: object
      title: SpotReadinessResponse
      description: Spot readiness response.
    SpotSweepPnlResponse:
      properties:
        type:
          anyOf:
          - type: string
          - type: 'null'
          title: Type
        status:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
        live_coinbase_orders_ran:
          type: boolean
          title: Live Coinbase Orders Ran
          default: false
      additionalProperties: true
      type: object
      title: SpotSweepPnlResponse
      description: Spot sweep P/L response.
    SpotSweepStatusResponse:
      properties:
        type:
          anyOf:
          - type: string
          - type: 'null'
          title: Type
        status:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
        live_coinbase_orders_ran:
          type: boolean
          title: Live Coinbase Orders Ran
          default: false
      additionalProperties: true
      type: object
      title: SpotSweepStatusResponse
      description: Spot sweep status response.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError